requirements.txt
pandas == 1.3.5
scikit_learn == 0.24.1
scikit_surprise == 1.1.1
scipy == 1.6.2
seaborn==0.12.2
openpyxl==3.1.2
